# openapi/relayfile-v1.openapi.yaml
openapi: 3.1.0
info:
  title: RelayFile API
  version: 1.0.0
  summary: Secure filesystem-over-REST contract for RelayFile
  description: |
    RelayFile exposes a workspace-scoped virtual filesystem with optimistic
    concurrency, async write-back operations, and event feeds.

    Security model:
    - Agent and service access uses Bearer JWT with workspace and scope claims.
    - Internal relay-cloud -> relayfile ingress uses HMAC-signed requests.
    - Correlation IDs are required for traceability and audit.
servers:
  - url: https://relayfile.agent-relay.com
    description: Production
  - url: https://relayfile.staging.agent-relay.com
    description: Staging
...
security:
  - BearerAuth: []
paths:
  /v1/workspaces/{workspaceId}/fs/tree:
    get:
      tags: [Filesystem]
      operationId: listTree
      summary: List filesystem entries under a path
      security:
        - BearerAuth: [fs:read]
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/CorrelationId'
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: /
            pattern: '^/.*'
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 2
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Tree listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workspaces/{workspaceId}/fs/file:
    get:
      tags: [Filesystem]
      operationId: readFile
      summary: Read file contents at a path
      security:
        - BearerAuth: [fs:read]
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/CorrelationId'
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: '^/.*'
      responses:
        '200':
          description: File read success
          headers:
            ETag:
              description: Current file revision identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileReadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Filesystem]
      operationId: writeFile
      summary: Write file content with optimistic concurrency
      security:
        - BearerAuth: [fs:write]
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IfMatch'
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: '^/.*'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileWriteRequest'
      responses:
        '202':
          description: Write accepted and queued for provider write-back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteQueuedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Filesystem]
      operationId: deleteFile
      summary: Delete a file with optimistic concurrency
      security:
        - BearerAuth: [fs:write]
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IfMatch'
        - name: path
          in: query
          required: true
          schema:
            type: string
            pattern: '^/.*'
      responses:
        '202':
          description: Delete accepted and queued for provider write-back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteQueuedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/workspaces/{workspaceId}/sync/status:
    get:
      tags: [Sync]
      operationId: getSyncStatus
      summary: Get provider sync health and cursor state for a workspace
      security:
        - BearerAuth: [sync:read]
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/CorrelationId'
        - name: provider
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Sync status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/workspaces/{workspaceId}/sync/refresh:
    post:
      tags: [Sync]
      operationId: triggerSyncRefresh
      summary: Trigger a provider refresh/reconciliation job
      security:
        - BearerAuth: [sync:trigger]
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRefreshRequest'
      responses:
        '202':
          description: Refresh job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/internal/webhook-envelopes:
    post:
      tags: [Internal]
      operationId: ingestWebhookEnvelope
      summary: Internal relay-cloud ingestion endpoint for normalized envelopes
      description: |
        Internal endpoint intended for relay-cloud only.
        Request must be HMAC signed and timestamp bounded to prevent replay.
      security:
        - InternalHmacAuth: []
      parameters:
        - $ref: '#/components/parameters/InternalTimestamp'
        - $ref: '#/components/parameters/InternalSignature'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEnvelopeRequest'
      responses:
        '202':
          description: Envelope accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Duplicate envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Ingress queue full, retry with backoff and jitter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/admin/replay/envelopes/{envelopeId}:
    post:
      tags: [Admin]
      operationId: replayEnvelope
      summary: Replay a previously ingested envelope from durable storage
      security:
        - BearerAuth: [admin:replay]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: envelopeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Replay accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/admin/replay/ops/{opId}:
    post:
      tags: [Admin]
      operationId: replayOperation
      summary: Replay a dead-lettered write-back operation
      security:
        - BearerAuth: [admin:replay]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: opId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Replay accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Operation is not replayable in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT must include workspace and agent claims.
        Required claims:
        - workspace_id
        - agent_name
        - scopes
        - exp
        - aud=relayfile
    InternalHmacAuth:
      type: apiKey
      in: header
      name: X-Relay-Signature
      description: |
        HMAC SHA-256 signature of canonical request body using shared internal secret.
        Must be accompanied by X-Relay-Timestamp and replay window checks.
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Required optimistic concurrency token (revision/etag).
...
  responses:
...
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    TreeResponse:
      type: object
      additionalProperties: false
      required: [path, entries]
      properties:
        path:
          type: string
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TreeEntry'
        nextCursor:
          type: string
          nullable: true
    FileReadResponse:
      type: object
      additionalProperties: false
      required: [path, revision, contentType, content]
      properties:
        path:
          type: string
        revision:
          type: string
        contentType:
          type: string
        content:
          type: string
        provider:
          type: string
        providerObjectId:
          type: string
        lastEditedAt:
          type: string
          format: date-time
...